pyproject.toml
[project]
name = "backend-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=7.0.5",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.42.0",
]
